+++ graylog2-web-interface/src/components/graylog/Accordion.tsx
@@ -14,69 +14,25 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-import PropTypes from 'prop-types';
 import React from 'react';
-import styled, { css } from 'styled-components';
+import PropTypes from 'prop-types';
 
-import Panel from './Panel';
+import { PanelGroup } from './bootstrap-import';
 
 type Props = {
   children: React.ReactNode,
-  name: string,
-  id?: string,
 }
 
-const StyledPanel = styled(Panel)`
-  border: 0;
-  border-radius: 0;
-`;
-
-const PanelHeading = styled(Panel.Heading)`
-  padding: 0;
-  border-radius: 0;
-`;
-
-const PanelTitle = styled(Panel.Title)(({ theme }) => css`
-  font-size: ${theme.fonts.size.small};
-  color: ${theme.colors.global.link};
-  
-  > a {
-    padding: 3px 9px;
-  }
-`);
-
-const PanelBody = styled(Panel.Body)(({ theme }) => css`
-  background-color: ${theme.colors.variant.lightest.default};
-  border-top-color: ${theme.colors.variant.light.default} !important;
-`);
-
-const Accordion = ({ children, name, id, ...restProps }: Props) => {
-  const eventKey = id ?? name.replace(/[^0-9a-zA-Z]/g, '-').toLowerCase();
-
+const Accordion = ({ children, ...restProps }:Props) => {
   return (
-    <StyledPanel {...restProps} id={id} eventKey={eventKey}>
-      <PanelHeading>
-        <PanelTitle toggle>
-          {name}
-        </PanelTitle>
-      </PanelHeading>
-      <Panel.Collapse>
-        <PanelBody>
-          {children}
-        </PanelBody>
-      </Panel.Collapse>
-    </StyledPanel>
+    <PanelGroup {...restProps} accordion>
+      {children}
+    </PanelGroup>
   );
 };
 
 Accordion.propTypes = {
-  name: PropTypes.node.isRequired,
-  id: PropTypes.string,
   children: PropTypes.node.isRequired,
-};
-
-Accordion.defaultProps = {
-  id: undefined,
 };
 
 export default Accordion;
